Hook into docusaurus blog plugin to remove compile step
Uses the maintainer's suggested solution to get blog posts for other
pages: https://github.com/facebook/docusaurus/discussions/6423#discussioncomment-2008482

Currently there's no better way than extending the blog plugin and
writing JSON files to disk. This allowed me to remove the compile step
and clean up the documentation. It'll make it easier to work with the
changed base URL.

To override the blog plugin, I had to disable the blog in the default
preset and move the config down, but it's the same as before.